requirements.txt
scikit-learn
statsmodels
flask
pymongo
certifi
python-dotenv
Flask_Cors
numpy
pandas
